Upgrade to xmlsectool version 1.1.1.  This version addresses the issue by which previous versions included unnecessary xmlns:xml prefix declarations:
https://bugs.internet2.edu/jira/browse/SECT-6

This fix means that it is no longer necessary to pipe xmlsectool's output through an identity transform to remove those declarations, so that step has been removed from the process as well.  This should speed the signing process up a little.  The only resulting change in output is one surplus xmlns:ds declaration on the signature, as described here:

https://bugs.internet2.edu/jira/browse/SECT-8

This is how metadatatool used to do things though, so it isn't a problem.