Rework approach to non-TLS mdui:Logo elements.
Although these are still rejected as errors for UKf registration, importing an mdui:Logo from an http:// URL causes the Logo element to be removed rather than causing a rejection. The result for the UKf is still that these are not republished to avoid mixed content errors, but an entity which *only* has this mistake is still imported albeit without its logo being available in discovery interfaces.