Externalise inputs and outputs into a shared workspace.
Inputs now come from entities.dir and members.dir, unsigned product
goes to output.dir and temporary files are put in temp.dir. The /xml/
directory which used to contain most of the above has been entirely
removed.

The shared workspace is defined by default in the dev environment, and
a non-dev environment will result in a failure if shared.ws.dir is not
defined via Jenkins.

A check is also made for the existence of the build subdirectory of the
shared workspace directory. This may be a step too far in the long run,
as many operations don't need it to be present.