Make compatible with Java 11
See ukf/ukf-meta#238.

This retains Xalan, but loads it from the classpath instead of the
endorsement mechanism used in earlier versions. This appears to work
(even with the extension classes) from the command line but might not
be stable in container environments, so we still need to move to
excluding Xerces and Xalan in favour of Java stages in the medium term.